Extract shared LLM timeout env parsing helpers
Move timeout env parsing and lazy fallback resolution into a single llm env helper module so provider adapters share identical validation behavior and error messages. This removes duplicated logic across Anthropic and OpenAI-compatible providers.
