+++ src/infra/llm/openai_compatible.rs
@@ -11,6 +11,7 @@ use crate::domain::{
 };
 
 use super::LlmProvider;
+use super::env::{read_env_var, read_timeout_from_env, resolve_timeout_with_global_fallback};
 use super::response_parsing::{extract_json_payload, truncate_message};
 use super::schema_validator::{GENERATION_RESULT_JSON_SCHEMA, LlmResponseSchemaValidator};
 
@@ -67,7 +68,12 @@ impl OpenAiCompatibleProvider {
             Some(value) => parse_supported_models(&value)?,
             None => default_supported_models(),
         };
-        let timeout = resolve_timeout_from_env()?;
+        let provider_timeout = read_timeout_from_env(ENV_TIMEOUT_SECS)?;
+        let timeout = resolve_timeout_with_global_fallback(
+            provider_timeout,
+            || read_timeout_from_env(ENV_GLOBAL_TIMEOUT_SECS),
+            DEFAULT_TIMEOUT,
+        )?;
 
         let mut provider = Self::with_config(
             provider_id,
@@ -593,56 +599,6 @@ fn parse_supported_models(value: &str) -> Result<Vec<String>, LlmError> {
     Ok(models)
 }
 
-fn parse_timeout_seconds(name: &str, value: &str) -> Result<Duration, LlmError> {
-    let parsed = value.trim().parse::<u64>().map_err(|_| {
-        LlmError::validation(format!("{name} must be a positive integer in seconds"))
-    })?;
-    if parsed == 0 {
-        return Err(LlmError::validation(format!(
-            "{name} must be greater than 0 seconds"
-        )));
-    }
-    Ok(Duration::from_secs(parsed))
-}
-
-fn read_timeout_from_env(name: &str) -> Result<Option<Duration>, LlmError> {
-    let Some(value) = read_env_var(name)? else {
-        return Ok(None);
-    };
-    Ok(Some(parse_timeout_seconds(name, &value)?))
-}
-
-fn resolve_timeout_from_env() -> Result<Duration, LlmError> {
-    let provider_timeout = read_timeout_from_env(ENV_TIMEOUT_SECS)?;
-    resolve_timeout_with_global_fallback(provider_timeout, || {
-        read_timeout_from_env(ENV_GLOBAL_TIMEOUT_SECS)
-    })
-}
-
-fn resolve_timeout_with_global_fallback<F>(
-    provider_timeout: Option<Duration>,
-    read_global_timeout: F,
-) -> Result<Duration, LlmError>
-where
-    F: FnOnce() -> Result<Option<Duration>, LlmError>,
-{
-    if let Some(timeout) = provider_timeout {
-        return Ok(timeout);
-    }
-
-    Ok(read_global_timeout()?.unwrap_or(DEFAULT_TIMEOUT))
-}
-
-fn read_env_var(name: &str) -> Result<Option<String>, LlmError> {
-    match std::env::var(name) {
-        Ok(value) => Ok(Some(value)),
-        Err(std::env::VarError::NotPresent) => Ok(None),
-        Err(error) => Err(LlmError::validation(format!(
-            "{name} could not be read: {error}"
-        ))),
-    }
-}
-
 fn read_bool_env(name: &str) -> Result<bool, LlmError> {
     let Some(value) = read_env_var(name)? else {
         return Ok(false);
@@ -676,17 +632,13 @@ fn build_v1_url(api_base_url: &str, endpoint_path: &str) -> String {
 
 #[cfg(test)]
 mod tests {
-    use super::{
-        OpenAiCompatibleProvider, build_v1_url, map_http_error, parse_bool, parse_timeout_seconds,
-        resolve_timeout_with_global_fallback,
-    };
+    use super::{OpenAiCompatibleProvider, build_v1_url, map_http_error, parse_bool};
     use crate::domain::{
         GenerationMode, GenerationParams, GenerationRequest, LlmError, MidiReferenceSummary,
         ModelRef, ReferenceSource,
     };
     use crate::infra::llm::LlmProvider;
     use reqwest::StatusCode;
-    use std::cell::Cell;
     use std::time::Duration;
 
     fn provider() -> OpenAiCompatibleProvider {
@@ -934,54 +886,5 @@ mod tests {
 
         let url = build_v1_url("https://example.com/v1/", "models");
         assert_eq!(url, "https://example.com/v1/models");
-    }
-
-    #[test]
-    fn parse_timeout_seconds_accepts_positive_integer_values() {
-        let timeout = parse_timeout_seconds("TEST_TIMEOUT", "8")
-            .expect("positive integer timeout should parse");
-        assert_eq!(timeout, Duration::from_secs(8));
-    }
-
-    #[test]
-    fn parse_timeout_seconds_rejects_invalid_values() {
-        let zero = parse_timeout_seconds("TEST_TIMEOUT", "0")
-            .expect_err("zero timeout should fail validation");
-        assert!(matches!(
-            zero,
-            LlmError::Validation { message }
-            if message == "TEST_TIMEOUT must be greater than 0 seconds"
-        ));
-
-        let invalid = parse_timeout_seconds("TEST_TIMEOUT", "abc")
-            .expect_err("non-integer timeout should fail validation");
-        assert!(matches!(
-            invalid,
-            LlmError::Validation { message }
-            if message == "TEST_TIMEOUT must be a positive integer in seconds"
-        ));
-    }
-
-    #[test]
-    fn resolve_timeout_with_global_fallback_is_lazy_for_provider_timeout() {
-        let global_called = Cell::new(false);
-
-        let timeout = resolve_timeout_with_global_fallback(Some(Duration::from_secs(3)), || {
-            global_called.set(true);
-            Err(LlmError::validation("global timeout should not be parsed"))
-        })
-        .expect("provider-specific timeout should short-circuit global fallback");
-
-        assert_eq!(timeout, Duration::from_secs(3));
-        assert!(!global_called.get());
-    }
-
-    #[test]
-    fn resolve_timeout_with_global_fallback_uses_global_when_provider_absent() {
-        let timeout =
-            resolve_timeout_with_global_fallback(None, || Ok(Some(Duration::from_secs(9))))
-                .expect("global timeout should be used when provider timeout is absent");
-
-        assert_eq!(timeout, Duration::from_secs(9));
     }
 }
